feat: finish relocating code editor to @srcbook/components
