+++ packages/web/package.json
@@ -13,8 +13,11 @@
     "check-types": "tsc"
   },
   "dependencies": {
+    "@codemirror/autocomplete": "^6.18.1",
     "@codemirror/lang-json": "^6.0.1",
     "@codemirror/lang-markdown": "^6.2.5",
+    "@codemirror/lang-javascript": "^6.2.2",
+    "@codemirror/lint": "^6.8.1",
     "@srcbook/shared": "workspace:^",
     "@srcbook/components": "workspace:^",
     "@uiw/codemirror-themes": "^4.23.2",
